// Web/RMA-Web-Shell/projects/medicare/src/app/preauth-manager/models/preauthorisation.ts
import{PreAuthorisationBreakdown} from 'projects/medicare/src/app/preauth-manager/models/preauthorisation-breakdown';
import{PreAuthIcd10Code} from 'projects/medicare/src/app/medi-manager/models/preAuthIcd10Code';
import { BaseClass } from 'projects/shared-models-lib/src/lib/common/base-class';
import { PreAuthTreatmentBasket } from 'projects/medicare/src/app/preauth-manager/models/preauth-treatment-basket';
import { PreAuthActivity } from 'projects/medicare/src/app/preauth-manager/models/preauth-activity';
import{PreAuthorisationUnderAssessReason} from 'projects/medicare/src/app/preauth-manager/models/preauthorisation-underassessreason';
import { PreAuthRehabilitation } from 'projects/medicare/src/app/preauth-manager/models/rehabilitation';
import { PreAuthMotivationForClaimReopening } from 'projects/medicare/src/app/preauth-manager/models/preauth-motivation-for-claim-reopening';
import { ChronicMedicationForm } from './chronic-medication-form';
import { ChronicMedicationFormRenewal } from './chronic-renewal-medication-form';
import { ProstheticQuotationTypeEnum } from '../../medi-manager/enums/prosthetic-quotation-type-enum';

export class PreAuthorisation extends BaseClass
{
  preAuthId: number;
  claimId: number
  personEventId: number;
  healthCareProviderId: number;
  requestingHealthCareProviderId: number;
  preAuthNumber: string;
  preAuthType: number;
  preAuthStatus: number;
  dateAuthorisedFrom: Date;
  dateAuthorisedTo: Date;
  dateAuthorised: Date;
  requestedAmount: number;
  authorisedAmount: number;
  requestComments: string;
  reviewComments: string;
  hospitalAuthId: number;
  isHighCost: boolean;
  isRequestFromHcp: boolean;
  temporaryReferenceNo: string;
  injuryDate: Date | null;
  isClaimLinked: boolean;
  isPatientVerified: boolean;
  preAuthContactNumber: string;
  isRehabilitationRequest: boolean;
  isWoundCareTreatment: boolean;
  isMedicationRequired: boolean;
  isClaimReopeningRequest: boolean;
  isPreRequest: boolean;
  preAuthorisationBreakdowns: PreAuthorisationBreakdown[];
  preAuthorisationUnderAssessReasons: PreAuthorisationUnderAssessReason[];
  preAuthRehabilitations: PreAuthRehabilitation[];
  preAuthIcd10Codes: PreAuthIcd10Code[];
  practitionerTypeId: number;
  subPreAuthorisations: PreAuthorisation[];
  preAuthTreatmentBaskets: PreAuthTreatmentBasket[];
  healthCareProviderName: string;
  practiceNumber: string;
  claimReferenceNumber: string;
  isVat: boolean;
  preAuthActivities: PreAuthActivity[];
  preAuthMotivationForClaimReopenings: PreAuthMotivationForClaimReopening[];
  isInHospital: boolean | null;
  prosthetistQuoteId: number | null;
  preAuthChronicRequestTypeId: number | null;
  scriptRepeats: string;
  comments: string;
  capturedDate: Date | null;
  preAuthChronicCMLs: number[];
  chronicMedicationForms: ChronicMedicationForm[];
  chronicMedicationFormRenewals: ChronicMedicationFormRenewal[];
  createdBy: string;
  chronicApplicationType: string;
  eventDate: Date | null;
  prostheticQuotationType: ProstheticQuotationTypeEnum | null;
  assignToUserId?: number;
  reviewWizardConfigId?: number;
}